Eventus ID 2.0
Eventus ID is a system for supervised coding of event data extracted
from text written in Spanish. The program is supervised as it requires
human intervention for developing dictionaries used as search categories
and for assessing the accuracy of the coding outcome. This software is
capable of coding event data based on a system of pattern recognition
that provides information on who did what to whom, when and where. The
flexibility provided by two event coding algorithms and an event locator
protocol fully integrated into the coding process enables Eventus ID to
identify fine-grained event data at the subnational level. To the best
of our knowledge, this is the first program specifically designed for
processing short news summaries written in Spanish.